Add ability to ignore caching
We need to be able to rebuild the redis cluster without causing errors
to our end users. We can set a maintenance mode in the runner but it
will still request the public key from the service token cache for any
request that it gets.

Setting the 'IGNORE_CACHE' environment variable will force public key to
be requested from where it is stored in the kubernetes cluster for every
request.

The rebuilding of the redis cluster does not take too long and we will
choose a time where there are lower numbers of users so it should not
impact services heavily.