Rescue CmdFailedError when kubectl command fails
In the instances of a request such as GET service/v3/foo/bar both the
service and the namespace do not exist. The kubectl command will fail
with a 500 and we raise a CmdFailedError. We want to rescue this and
actually return and empty string which we translate into a 404 for the
requesting service.

We feel this is more accurate for what is actually happening. We only
rescue the CmdFailedError so an actual 500 error will still respond
correctly.

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>
Co-authored-by: Matt Tei <matt.tei@digital.justice.gov.uk>
Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>