Cater for ignore_cache query param
We need to be able to cater for rotating the public keys. The service
token cache has the `IGNORE_CACHE` environment variable which we could
set to 'true' and then deploy the app out. However that would mean that
every request received for every app would hit Kubernetes.

Passing an `ignore_cache` query parameter allows us to be more flexible
when requesting the public key from Kubernetes. We only need to do so
for the specific app that needs it instead of all of them at the same
time