Use service_slug for consistency
The same property is referenced across the platform using several
different things but it always represents the same thing.

Eventually the actual thing that the request to the service token cache
is trying to retrieve is the public key from Kubernetes or from the
Redis if it has already been cached there.

The service slug is a parameterized version of the service name. But
with version 3 of the service token cache endpoint the service_slug
could also be related to a platform app, not just a form.

In Redis the key which is used takes the form:

`encoded-public-key-<service_slug>`

e.g `encoded-public-key-apply-financial-deputy`

And then if the cache has expired then the request to Kubernetes will
look for a config map, _not_ a secret, which takes the form:

`fb-<service_slug>-config-map`

e.g

`fb-apply-financial-deputy-config-map`

The name of this config map is created during the deployment of the
platform application or the service so it has to match that.

Regardless, service_slug seems like a more universal description as
opposed to application and definitely is a lot more accurate than
secret_name as at no point does the service token cache ever request a
kubernetes secret.